Divide
클래스 나눔
