+++ TeamTextRPG/Start.cs
@@ -3,9 +3,8 @@ using System.Numerics;
 using System.Text;
 using static System.Net.Mime.MediaTypeNames;
 
-namespace TextRPGGame
+namespace TeamTextRPG
 {
-    // Console.BackgroundColor = ConsoleColor.Red; 데미지 입었을 때 화면이 빨간색으로 변하면 재밌겠다
     internal class Start
     {
         private static Character player;
@@ -59,7 +58,7 @@ namespace TextRPGGame
             Console.WriteLine("          ===== 계속하려면 아무 키나 입력하십시오. =====          ");
             Console.ResetColor();
             Console.ReadKey();
-            Console.Beep();
+            Console.Beep(); // 삡 소리 나게 하는 것
         }
 
         // 스타트
@@ -234,6 +233,7 @@ namespace TextRPGGame
             items[index].IsEquipped = !items[index].IsEquipped; // ! : 불형의 변수를 반대로 만들어주는 것
         }
 
+        // 상속? Shop부분에 가져와야함
         // 상점
         static void DisplayShop()
         {
@@ -292,11 +292,11 @@ namespace TextRPGGame
             {
                 if (Console.KeyAvailable) // 아무 키나 눌렸을 때 true
                 {
-                    speed = 0;
+                    speed = 0; // 아무 키나 눌렀을 때 빠르게 즉, 전체 문장이 나오도록 skip하는 기능
                     Console.ReadKey(true);
                 }
                 Console.Write(txt[txtCount]); // 글자를 하나하나 가져올 인덱스
-                Thread.Sleep(speed);
+                Thread.Sleep(speed); // 속도 지정 메서드
                 txtCount++;
             }
         }
@@ -328,117 +328,122 @@ namespace TextRPGGame
         }
     }
 
+    // 캐릭터
+    #region
+    //public class Character
+    //{
+    //    public string Name { get; }
+    //    public string Job { get; }
+    //    public int Level { get; }
+    //    public int Atk { get; }
+    //    public int Def { get; }
+    //    public int Hp { get; }
+    //    public int Gold { get; }
 
-    public class Character
-    {
-        public string Name { get; }
-        public string Job { get; }
-        public int Level { get; }
-        public int Atk { get; }
-        public int Def { get; }
-        public int Hp { get; }
-        public int Gold { get; }
+    //    public Character(string name, string job, int level, int atk, int def, int hp, int gold)
+    //    {
+    //        Name = name;
+    //        Job = job;
+    //        Level = level;
+    //        Atk = atk;
+    //        Def = def;
+    //        Hp = hp;
+    //        Gold = gold;
+    //    }
+    //}
+    #endregion
 
-        public Character(string name, string job, int level, int atk, int def, int hp, int gold)
-        {
-            Name = name;
-            Job = job;
-            Level = level;
-            Atk = atk;
-            Def = def;
-            Hp = hp;
-            Gold = gold;
-        }
-    }
+    // 아이템
+    #region
+    //public class Item
+    //{
+    //    // 장비 타입
+    //    public enum ItemType
+    //    {
+    //        Weapon,
+    //        Head,
+    //        Body,
+    //        Shoes,
+    //        Acc
+    //    }
 
-    public class Item
-    {
-        // 장비 타입
-        public enum ItemType
-        {
-            Weapon,
-            Head,
-            Body,
-            Shoes,
-            Acc
-        }
+    //    public string Name { get; }
+    //    public string Description { get; }
+    //    public ItemType Type { get; }
+    //    public int AtkOption { get; }
+    //    public int DefOption { get; }
+    //    public int HpOption { get; }
+    //    public bool IsEquipped { get; set; }
 
-        public string Name { get; }
-        public string Description { get; }
-        public ItemType Type { get; }
-        public int AtkOption { get; }
-        public int DefOption { get; }
-        public int HpOption { get; }
-        public bool IsEquipped { get; set; }
+    //    public static int ItemCount = 0; // static을 붙임으로 Item이라는 클래스에 귀속된다
 
-        public static int ItemCount = 0; // static을 붙임으로 Item이라는 클래스에 귀속된다
+    //    public Item(string name, string description, ItemType type, int atkOption, int defOption, int hpOption, bool isEquipped = false)
+    //    {
+    //        Name = name;
+    //        Description = description;
+    //        Type = type;
+    //        AtkOption = atkOption;
+    //        DefOption = defOption;
+    //        HpOption = hpOption;
+    //        IsEquipped = isEquipped;
+    //    }
 
-        public Item(string name, string description, ItemType type, int atkOption, int defOption, int hpOption, bool isEquipped = false)
-        {
-            Name = name;
-            Description = description;
-            Type = type;
-            AtkOption = atkOption;
-            DefOption = defOption;
-            HpOption = hpOption;
-            IsEquipped = isEquipped;
-        }
+    //    // 장비 착용 표시
+    //    public void ItemStat(bool equippedItem = false, int index = 0) // 기본값이 false
+    //    {
+    //        Console.Write("- ");
+    //        if (equippedItem)
+    //        {
+    //            Console.ForegroundColor = ConsoleColor.Blue;
+    //            Console.Write("{0} ", index); // 몇 번째 아이템인가
+    //            Console.ResetColor();
+    //        }
+    //        if (IsEquipped)
+    //        {
+    //            Console.Write("[");
+    //            Console.ForegroundColor = ConsoleColor.Magenta;
+    //            Console.Write("E");
+    //            Console.ResetColor();
+    //            Console.Write("]");
+    //            Console.WriteLine(PadRightForMixedText(Name, 9)); // 방어력, 공격력 옵션들 간격맞춤
+    //        }
+    //        else
+    //            Console.WriteLine(PadRightForMixedText(Name, 12));
+    //        Console.Write(" | ");
 
-        // 장비 착용 표시
-        public void ItemStat(bool equippedItem = false, int index = 0) // 기본값이 false
-        {
-            Console.Write("- ");
-            if (equippedItem)
-            {
-                Console.ForegroundColor = ConsoleColor.Blue;
-                Console.Write("{0} ", index); // 몇 번째 아이템인가
-                Console.ResetColor();
-            }
-            if (IsEquipped)
-            {
-                Console.Write("[");
-                Console.ForegroundColor = ConsoleColor.Magenta;
-                Console.Write("E");
-                Console.ResetColor();
-                Console.Write("]");
-                Console.WriteLine(PadRightForMixedText(Name, 9)); // 방어력, 공격력 옵션들 간격맞춤
-            }
-            else
-                Console.WriteLine(PadRightForMixedText(Name, 12));
-            Console.Write(" | ");
+    //        if (AtkOption != 0) Console.Write($"공격력 {(AtkOption >= 0 ? "+" : "")}{AtkOption} "); // 삼항연산자
+    //        if (DefOption != 0) Console.Write($"방어력 {(DefOption >= 0 ? "+" : "")}{DefOption} "); // 옵션이 0이 아니라면 옵션수치를 내보내라
+    //        if (HpOption != 0) Console.Write($"체 력 {(HpOption >= 0 ? "+" : "")}{HpOption} "); // [조건 ? 참 : 거짓]
 
-            if (AtkOption != 0) Console.Write($"공격력 {(AtkOption >= 0 ? "+" : "")}{AtkOption} "); // 삼항연산자
-            if (DefOption != 0) Console.Write($"방어력 {(DefOption >= 0 ? "+" : "")}{DefOption} "); // 옵션이 0이 아니라면 옵션수치를 내보내라
-            if (HpOption != 0) Console.Write($"체 력 {(HpOption >= 0 ? "+" : "")}{HpOption} "); // [조건 ? 참 : 거짓]
+    //        Console.Write(" | ");
 
-            Console.Write(" | ");
+    //        Console.WriteLine(Description);
+    //    }
 
-            Console.WriteLine(Description);
-        }
+    //    // 정렬
+    //    public static int GetPrintableLength(string str)
+    //    {
+    //        int length = 0;
+    //        foreach (char c in str)
+    //        {
+    //            if (char.GetUnicodeCategory(c) == System.Globalization.UnicodeCategory.OtherLetter)
+    //            {
+    //                length += 2; // 한글과 같은 넓은 문자에 대해 길이를 2로 취급
+    //            }
+    //            else
+    //            {
+    //                length += 1; // 나머지 문자에 대해 길이를 1로 취급
+    //            }
+    //        }
 
-        // 정렬
-        public static int GetPrintableLength(string str)
-        {
-            int length = 0;
-            foreach (char c in str)
-            {
-                if (char.GetUnicodeCategory(c) == System.Globalization.UnicodeCategory.OtherLetter)
-                {
-                    length += 2; // 한글과 같은 넓은 문자에 대해 길이를 2로 취급
-                }
-                else
-                {
-                    length += 1; // 나머지 문자에 대해 길이를 1로 취급
-                }
-            }
-
-            return length;
-        }
-        public static string PadRightForMixedText(string str, int totalLength)
-        {
-            int currentLength = GetPrintableLength(str); // 텍스트의 길이를 구한다
-            int padding = totalLength - currentLength; // 원하는 크기과 실제 텍스트의 길이를 계산한다
-            return str.PadRight(str.Length + padding); // 필요한 길이만큼의 공간을 더해준다
-        }
-    }
+    //        return length;
+    //    }
+    //    public static string PadRightForMixedText(string str, int totalLength)
+    //    {
+    //        int currentLength = GetPrintableLength(str); // 텍스트의 길이를 구한다
+    //        int padding = totalLength - currentLength; // 원하는 크기과 실제 텍스트의 길이를 계산한다
+    //        return str.PadRight(str.Length + padding); // 필요한 길이만큼의 공간을 더해준다
+    //    }
+    //}
+    #endregion
 }
